feat(simple-mission): allow a memo to be left when a mission is fulfilled
The memo may simply help to link off-chain work with fulfilment of a mission. For a simple mission, it would be a voluntary provision of extra reference to the related work by the mission operator. Similar to the reference and particular fields for a payment in everyday banking. At this level and for simple contracts we would be agnostic about what should go in that field.